requirements.txt
bpy==4.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
Cython==3.0.11
idna==3.7
numpy==2.1.0
pillow==10.4.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
rectpack==0.2.2
requests==2.32.3
urllib3==2.2.2
zstandard==0.23.0
